modularize provisioner: introduce local STS account service
