+++ core/identity-hub-participants/src/main/java/org/eclipse/edc/identityhub/participantcontext/ParticipantContextExtension.java
@@ -16,22 +16,20 @@ package org.eclipse.edc.identityhub.participantcontext;
 
 import org.eclipse.edc.identithub.spi.did.store.DidResourceStore;
 import org.eclipse.edc.identityhub.spi.keypair.KeyPairService;
-import org.eclipse.edc.identityhub.spi.participantcontext.AccountProvisioner;
 import org.eclipse.edc.identityhub.spi.participantcontext.ParticipantContextService;
+import org.eclipse.edc.identityhub.spi.participantcontext.StsAccountProvisioner;
 import org.eclipse.edc.identityhub.spi.participantcontext.events.ParticipantContextObservable;
 import org.eclipse.edc.identityhub.spi.store.ParticipantContextStore;
 import org.eclipse.edc.runtime.metamodel.annotation.Extension;
 import org.eclipse.edc.runtime.metamodel.annotation.Inject;
 import org.eclipse.edc.runtime.metamodel.annotation.Provider;
 import org.eclipse.edc.spi.event.EventRouter;
-import org.eclipse.edc.spi.result.ServiceResult;
 import org.eclipse.edc.spi.security.Vault;
 import org.eclipse.edc.spi.system.ServiceExtension;
 import org.eclipse.edc.transaction.spi.TransactionContext;
 
 import java.time.Clock;
 
-import static java.util.Optional.ofNullable;
 import static org.eclipse.edc.identityhub.participantcontext.ParticipantContextExtension.NAME;
 
 @Extension(NAME)
@@ -53,8 +51,8 @@ public class ParticipantContextExtension implements ServiceExtension {
     @Inject
     private DidResourceStore didResourceStore;
 
-    @Inject(required = false)
-    private AccountProvisioner accountProvisioner;
+    @Inject
+    private StsAccountProvisioner stsAccountProvisioner;
 
     private ParticipantContextObservable participantContextObservable;
 
@@ -65,7 +63,7 @@ public class ParticipantContextExtension implements ServiceExtension {
 
     @Provider
     public ParticipantContextService createParticipantService() {
-        return new ParticipantContextServiceImpl(participantContextStore, didResourceStore, vault, transactionContext, participantContextObservable(), accountProvisioner());
+        return new ParticipantContextServiceImpl(participantContextStore, didResourceStore, vault, transactionContext, participantContextObservable(), stsAccountProvisioner);
     }
 
     @Provider
@@ -75,10 +73,5 @@ public class ParticipantContextExtension implements ServiceExtension {
             participantContextObservable.registerListener(new ParticipantContextEventPublisher(clock, eventRouter));
         }
         return participantContextObservable;
-    }
-
-    private AccountProvisioner accountProvisioner() {
-        return ofNullable(accountProvisioner)
-                .orElseGet(() -> manifest -> ServiceResult.success()); // default is a NOOP
     }
 }
