+++ extensions/sts/sts-account-provisioner/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
