+++ extensions/sts/sts-account-service-local/build.gradle.kts
