+++ launcher/build.gradle.kts
@@ -26,7 +26,8 @@ dependencies {
     runtimeOnly(project(":core:identity-hub-keypairs"))
     runtimeOnly(project(":extensions:did:local-did-publisher"))
     runtimeOnly(project(":extensions:common:credential-watchdog"))
-    runtimeOnly(project(":extensions:common:sts-account-provisioner"))
+    runtimeOnly(project(":extensions:sts:sts-account-provisioner"))
+    runtimeOnly(project(":extensions:sts:sts-account-service-local"))
     runtimeOnly(project(":extensions:api:identity-api:did-api"))
     runtimeOnly(project(":extensions:api:identity-api:participant-context-api"))
     runtimeOnly(project(":extensions:api:identity-api:verifiable-credentials-api"))
