+++ settings.gradle.kts
@@ -50,7 +50,9 @@ include(":extensions:store:sql:identity-hub-participantcontext-store-sql")
 include(":extensions:store:sql:identity-hub-keypair-store-sql")
 include(":extensions:did:local-did-publisher")
 include(":extensions:common:credential-watchdog")
-include(":extensions:common:sts-account-provisioner")
+include(":extensions:sts:sts-account-provisioner")
+include(":extensions:sts:sts-account-service-local")
+include(":extensions:sts:sts-account-service-remote")
 
 // Identity APIs
 include(":extensions:api:identity-api:validators")
