+++ spi/participant-context-spi/build.gradle.kts
@@ -22,5 +22,6 @@ val swagger: String by project
 
 dependencies {
     implementation(libs.edc.spi.identity.did) // ParticipantManifest#serviceEndpoint
+    implementation(libs.edc.sts.spi) // StsAccount
     testImplementation(libs.edc.lib.json)
 }
